requirement.txt
click==8.0.4
dynaconf==3.1.7
Flask==2.0.3
Flask-DebugToolbar==0.11.0
gunicorn==20.1.0
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.1
PyYAML==6.0
waitress==2.1.2
Werkzeug==2.2.3
gmplot==1.4.1